+++ src/MusicDatabase1.java
@@ -1,6 +1,7 @@
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.Iterator;
+import java.util.NoSuchElementException;
 
 /**
  * {@code MusicDatabase} represented as an {@link java.util.ArrayList
@@ -59,9 +60,31 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
     @Override
     public Iterator<Song> iterator() {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'iterator'");
+        Iterator<Song> it = new Iterator<Song>() {
+            private int pos = 0;
+
+            @Override
+            public boolean hasNext() {
+                Boolean hasNextValue = false;
+                if (this.pos < MusicDatabase1.this.db.size()
+                        && MusicDatabase1.this.db.get(this.pos) != null) {
+                    hasNextValue = true;
+                }
+                return hasNextValue;
+            }
+
+            @Override
+            public Song next() {
+                if (!this.hasNext()) {
+                    throw new NoSuchElementException(
+                            "ERROR: No more values in this");
+                }
+                Song value = MusicDatabase1.this.db.get(this.pos);
+                this.pos++;
+                return value;
+            }
+        };
+        return it;
     }
 
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
@@ -103,19 +126,63 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
 
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
     @Override
-    public Song getEntryInOrder(int n) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'getEntryInOrder'");
+    public Song getEntryByOrder(int n) {
+        /*
+         * ArrayList's get method will already return an error if n is out of
+         * bounds, so so this method only needs to throw an exception if an
+         * element doesn't exist.
+         */
+        Song entry = this.db.get(n);
+
+        if (entry == null) {
+            throw new NoSuchElementException(
+                    "ERROR: No element exists at the given position");
+        }
+
+        return entry;
     }
 
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
     @Override
     public ArrayList<Song> getEntries(MusicDatabaseKernel.SearchField field,
             String value) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'getEntries'");
+        assert value != null : "Violation of: value != null";
+
+        ArrayList<Song> list = new ArrayList<Song>();
+
+        for (Song song : this.db) {
+            switch (field) {
+                case TITLE: {
+                    if (song.title().equals(value)) {
+                        list.add(song);
+                    }
+                    break;
+                }
+                case ARTIST: {
+                    if (song.artist().equals(value)) {
+                        list.add(song);
+                    }
+                    break;
+                }
+                case ALBUM: {
+                    if (song.album().equals(value)) {
+                        list.add(song);
+                    }
+                    break;
+                }
+                case LENGTH: {
+                    if (song.length().equals(value)) {
+                        list.add(song);
+                    }
+                    break;
+                }
+                default: {
+                    break;
+                }
+            }
+        }
+
+        return list;
     }
 
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
@@ -134,6 +201,24 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
 
         this.db.remove(song);
         return song;
+    }
+
+    // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
+    @Override
+    public Song removeEntryByOrder(int n) {
+        /*
+         * ArrayList's get method will already return an error if n is out of
+         * bounds, so so this method only needs to throw an exception if an
+         * element doesn't exist.
+         */
+        Song entry = this.db.get(n);
+
+        if (entry == null) {
+            throw new NoSuchElementException(
+                    "ERROR: No element exists at the given position");
+        }
+
+        return this.db.remove(n);
     }
 
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN