+++ src/MusicDatabaseKernel.java
@@ -13,7 +13,7 @@ public interface MusicDatabaseKernel
     /**
      * An enum for methods only intended to accept the four fields of the
      * {@code Song} class as one of their parameters (i.e. methods searching by
-     * field).
+     * field). Possible values: TITLE, ARTIST, ALBUM, LENGTH
      */
     enum SearchField {
         /**
@@ -36,26 +36,29 @@ public interface MusicDatabaseKernel
     void addEntry(Song song);
 
     /**
-     * Returns the {@code Song} at position {@code n} in the database. Returns
-     * -1 if no object exists at that position.
+     * Returns the {@code Song} at position {@code n} in the database. Throws a
+     * {@code NoSuchElementException} if there is no Song object at the
+     * position.
      *
      * @param n
      *            An int representing the index at which the user wants to fetch
      *            a value from.
      * @return A {@code Song} if one exists at {@code n}.
-     * @ensures getEntryInOrder = The {@code Song} at position {@code n} if it
-     *          exists, or -1 otherwise
+     * @requires A {@code Song} object exists at index n of this
+     * @ensures getEntryByOrder = The {@code Song} at position {@code n} if it
+     *          exists
      */
-    Song getEntryInOrder(int n);
+    Song getEntryByOrder(int n);
 
     /**
-     * A method that takes an enum (which can be "title", "artist", "album", or
-     * "length"), and a value to use to find matches within the {@code Song}
+     * A method that takes an enum (which can be TITLE, ARTIST, ALBUM, or
+     * LENGTH), and a value to use to find matches within the {@code Song}
      * objects of the database. It then returns an {@code ArrayList} of all
      * matches found.
      *
      * @param field
-     *            An enum representing the {@code Song} field to search through
+     *            An enum (possible values: TITLE, ARTIST, ALBUM, LENGTH)
+     *            representing the {@code Song} field to search through
      * @param value
      *            A string value to search by
      * @return An {@code ArrayList} of all {@code Songs} with a field matching
@@ -97,6 +100,19 @@ public interface MusicDatabaseKernel
      * @updates this
      */
     Song removeEntry(Song song);
+
+    /**
+     * Takes in an index and removes the {@code Song} there from the database.
+     * Returns an error if no entry exists at n.
+     *
+     * @param n
+     *            The index of a song to be removed
+     * @return The {@code Song} removed.
+     * @requires An entry exists at n.
+     * @ensures this = #this / [song at n]
+     * @updates this
+     */
+    Song removeEntryByOrder(int n);
 
     /**
      * A method that takes in a Song object and checks for a match in the